Delhi ideas buffed
They are so bad holy shit, the indian sultanate ideas are better with 10% cca / 10 cav cost tradition and 10 ICA as finisher that delhi does not have. Adding it so delhi ideas are a total joke. They are still pretty bad even with this being added and it's still the weakest set in india, but no longer that bad.